+++ lib/layrz_icons.dart
@@ -2,7 +2,8 @@ library layrz_icons;
 
 import 'package:flutter/widgets.dart';
 
-part 'src/enum.dart';
+part 'src/class_enum.dart';
+part 'src/icon_enum.dart';
 part 'src/mapping.dart';
 
 class LayrzIcon {
@@ -33,6 +34,10 @@ enum LayrzFamily {
   fontAwesomeSharpLight,
   fontAwesomeSharpRegular,
   fontAwesomeSharpSolid,
+  ionicons,
+  iconsaxPlusBold,
+  iconsaxPlusBroken,
+  iconsaxPlusLinear,
   ;
 
   String get fontFamily {
@@ -63,6 +68,14 @@ enum LayrzFamily {
         return 'FontAwesomeSharpRegular';
       case LayrzFamily.fontAwesomeSharpSolid:
         return 'FontAwesomeSharpSolid';
+      case LayrzFamily.ionicons:
+        return 'Ionicons';
+      case LayrzFamily.iconsaxPlusBold:
+        return 'IconsaxPlusBold';
+      case LayrzFamily.iconsaxPlusBroken:
+        return 'IconsaxPlusBroken';
+      case LayrzFamily.iconsaxPlusLinear:
+        return 'IconsaxPlusLinear';
     }
   }
 
@@ -84,6 +97,12 @@ enum LayrzFamily {
       case LayrzFamily.fontAwesomeSharpRegular:
       case LayrzFamily.fontAwesomeSharpSolid:
         return 'font_awesome_flutter';
+      case LayrzFamily.ionicons:
+        return 'ionicons';
+      case LayrzFamily.iconsaxPlusBold:
+      case LayrzFamily.iconsaxPlusBroken:
+      case LayrzFamily.iconsaxPlusLinear:
+        return 'iconsax_plus';
     }
   }
 }