+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Changelog
 
+## 1.0.3
+
+- Optimizations
+- Added `ionicons` to the integrated icons
+- Added `iconsax_plus` to the integrated icons
+
+## 1.0.2
+
+- Added `LayrzIconsClasses` to simplify the class calling.
+
 ## 1.0.1
 
-* Added `homepage` on `pubspec.yaml`
-* Updated `README.md`
+- Added `homepage` on `pubspec.yaml`
+- Updated `README.md`
 
 ## 1.0.0
 
-* Initial release
+- Initial release